+++ dimpy/input_file/__init__.py
@@ -1,0 +1,1 @@
+from .read_input_file import ReadInput