+++ dimpy/methods/__init__.py
@@ -1,0 +1,6 @@
+from .base import CalcMethodBase
+from .static_dda import DDAs
+from .static_dim import DIMs
+from .static_dda_pbc import DDAsPBC
+from .dynamic_dda import DDAr
+from .dynamic_dda_pbc import DDArPBC