+++ dimpy/nanoparticle/__init__.py
@@ -1,0 +1,1 @@
+from .nanoparticle import Nanoparticle