+++ CHANGELOG.rst
@@ -8,11 +8,26 @@ All notable changes to this project will be documented in this file.
 Unreleased
 ##########
 
-- DIM
 - DIM with retardation
 - DIM with PBC
 - DIM with PBC and retardation
 - DIM/QM with PySCF
+
+[0.6.0] - 2021-06-29
+####################
+
+Added
+-----
+
+- DIM
+- Analyzer
+
+Changed
+-------
+
+- analyzer, input_file, methods, nanoparticle, tools modules are now
+  their own subpackages of DIMPy
+
 
 [0.5.0] - 2021-06-18
 ####################